+++ src/flask_sqlalchemy/query.py
@@ -1,5 +1,4 @@
 from flask import abort
-from flask import request
 from sqlalchemy import orm
 
 from .pagination import Pagination
@@ -34,77 +33,44 @@ class BaseQuery(orm.Query):
         return rv
 
     def paginate(
-        self, page=None, per_page=None, error_out=True, max_per_page=None, count=True
-    ):
-        """Returns ``per_page`` items from page ``page``.
+        self,
+        *,
+        page: int | None = None,
+        per_page: int | None = None,
+        max_per_page: int | None = None,
+        error_out: bool = True,
+        count: bool = True,
+    ) -> Pagination:
+        """Apply an offset and limit to the query based on the current page and number
+        of items per page, returning a :class:`.Pagination` object.
 
-        If ``page`` or ``per_page`` are ``None``, they will be retrieved from
-        the request query. If ``max_per_page`` is specified, ``per_page`` will
-        be limited to that value. If there is no request or they aren't in the
-        query, they default to 1 and 20 respectively. If ``count`` is ``False``,
-        no query to help determine total page count will be run.
+        :param query: The query to paginate.
+        :param page: The current page, used to calculate the offset. Defaults to the
+            ``page`` query arg during a request, or 1 otherwise.
+        :param per_page: The maximum number of items on a page, used to calculate the
+            offset and limit. Defaults to the ``per_page`` query arg during a request,
+            or 20 otherwise.
+        :param max_per_page: The maximum allowed value for ``per_page``, to limit a
+            user-provided value.
+        :param error_out: Abort with a ``404 Not Found`` error if no items are returned
+            and ``page`` is not 1, or if ``page`` is less than 1 or ``per_page`` is
+            negative, or if either are not ints. If disabled, an invalid ``page``
+            defaults to 1, and ``per_page`` defaults to 20.
+        :param count: Calculate the total number of values by issuing an extra count
+            query. For very complex queries this may be inaccurate or slow, so it can be
+            disabled and set manually if necessary.
 
-        When ``error_out`` is ``True`` (default), the following rules will
-        cause a 404 response:
+        .. versionchanged:: 3.0
+            All parameters are keyword-only.
 
-        * No items are found and ``page`` is not 1.
-        * ``page`` is less than 1, or ``per_page`` is negative.
-        * ``page`` or ``per_page`` are not ints.
-
-        When ``error_out`` is ``False``, ``page`` and ``per_page`` default to
-        1 and 20 respectively.
-
-        Returns a :class:`Pagination` object.
+        .. versionchanged:: 3.0
+            The ``count`` query is more efficient.
         """
-
-        if request:
-            if page is None:
-                try:
-                    page = int(request.args.get("page", 1))
-                except (TypeError, ValueError):
-                    if error_out:
-                        abort(404)
-
-                    page = 1
-
-            if per_page is None:
-                try:
-                    per_page = int(request.args.get("per_page", 20))
-                except (TypeError, ValueError):
-                    if error_out:
-                        abort(404)
-
-                    per_page = 20
-        else:
-            if page is None:
-                page = 1
-
-            if per_page is None:
-                per_page = 20
-
-        if max_per_page is not None:
-            per_page = min(per_page, max_per_page)
-
-        if page < 1:
-            if error_out:
-                abort(404)
-            else:
-                page = 1
-
-        if per_page < 0:
-            if error_out:
-                abort(404)
-            else:
-                per_page = 20
-
-        items = self.limit(per_page).offset((page - 1) * per_page).all()
-
-        if not items and page != 1 and error_out:
-            abort(404)
-
-        if not count:
-            total = None
-        else:
-            total = self.order_by(None).count()
-
-        return Pagination(self, page, per_page, total, items)
+        return Pagination.apply_to_query(
+            self,
+            page=page,
+            per_page=per_page,
+            max_per_page=max_per_page,
+            error_out=error_out,
+            count=count,
+        )